requirements.txt
fastapi==0.88.0
requests
pillow